+++ src/main/java/edu/ie3/datamodel/io/factory/result/ConnectorResultFactory.java
@@ -7,8 +7,8 @@ package edu.ie3.datamodel.io.factory.result;
 
 import edu.ie3.datamodel.exceptions.FactoryException;
 import edu.ie3.datamodel.io.factory.EntityData;
+import edu.ie3.datamodel.models.Entity;
 import edu.ie3.datamodel.models.StandardUnits;
-import edu.ie3.datamodel.models.UniqueEntity;
 import edu.ie3.datamodel.models.result.connector.ConnectorResult;
 import edu.ie3.datamodel.models.result.connector.LineResult;
 import edu.ie3.datamodel.models.result.connector.Transformer2WResult;
@@ -49,23 +49,20 @@ public class ConnectorResultFactory extends ResultEntityFactory<ConnectorResult>
   protected List<Set<String>> getFields(Class<?> entityClass) {
     /// all result models have the same constructor except StorageResult
     Set<String> minConstructorParams = newSet(TIME, INPUT_MODEL, IAMAG, IAANG, IBMAG, IBANG);
-    Set<String> optionalFields = expandSet(minConstructorParams, ENTITY_UUID);
 
     if (entityClass.equals(Transformer2WResult.class)) {
       minConstructorParams = newSet(TIME, INPUT_MODEL, IAMAG, IAANG, IBMAG, IBANG, TAPPOS);
-      optionalFields = expandSet(minConstructorParams, ENTITY_UUID);
     } else if (entityClass.equals(Transformer3WResult.class)) {
       minConstructorParams =
           newSet(TIME, INPUT_MODEL, IAMAG, IAANG, IBMAG, IBANG, ICMAG, ICANG, TAPPOS);
-      optionalFields = expandSet(minConstructorParams, ENTITY_UUID);
     }
 
-    return Arrays.asList(minConstructorParams, optionalFields);
+    return List.of(minConstructorParams);
   }
 
   @Override
   protected ConnectorResult buildModel(EntityData data) {
-    final Class<? extends UniqueEntity> entityClass = data.getTargetClass();
+    final Class<? extends Entity> entityClass = data.getTargetClass();
     ZonedDateTime time = timeUtil.toZonedDateTime(data.getField(TIME));
 
     UUID inputModel = data.getUUID(INPUT_MODEL);
@@ -76,23 +73,12 @@ public class ConnectorResultFactory extends ResultEntityFactory<ConnectorResult>
         data.getQuantity(IBMAG, StandardUnits.ELECTRIC_CURRENT_MAGNITUDE);
     ComparableQuantity<Angle> iBAng = data.getQuantity(IBANG, StandardUnits.ELECTRIC_CURRENT_ANGLE);
 
-    Optional<UUID> uuidOpt =
-        data.containsKey(ENTITY_UUID) ? Optional.of(data.getUUID(ENTITY_UUID)) : Optional.empty();
-
     if (entityClass.equals(LineResult.class))
-      return uuidOpt
-          .map(uuid -> new LineResult(uuid, time, inputModel, iAMag, iAAng, iBMag, iBAng))
-          .orElseGet(() -> new LineResult(time, inputModel, iAMag, iAAng, iBMag, iBAng));
+      return new LineResult(time, inputModel, iAMag, iAAng, iBMag, iBAng);
     else if (entityClass.equals(Transformer2WResult.class)) {
       final int tapPos = data.getInt(TAPPOS);
 
-      return uuidOpt
-          .map(
-              uuid ->
-                  new Transformer2WResult(
-                      uuid, time, inputModel, iAMag, iAAng, iBMag, iBAng, tapPos))
-          .orElseGet(
-              () -> new Transformer2WResult(time, inputModel, iAMag, iAAng, iBMag, iBAng, tapPos));
+      return new Transformer2WResult(time, inputModel, iAMag, iAAng, iBMag, iBAng, tapPos);
     } else if (entityClass.equals(Transformer3WResult.class)) {
       ComparableQuantity<ElectricCurrent> iCMag =
           data.getQuantity(ICMAG, StandardUnits.ELECTRIC_CURRENT_MAGNITUDE);
@@ -100,15 +86,8 @@ public class ConnectorResultFactory extends ResultEntityFactory<ConnectorResult>
           data.getQuantity(ICANG, StandardUnits.ELECTRIC_CURRENT_ANGLE);
       final int tapPos = data.getInt(TAPPOS);
 
-      return uuidOpt
-          .map(
-              uuid ->
-                  new Transformer3WResult(
-                      uuid, time, inputModel, iAMag, iAAng, iBMag, iBAng, iCMag, iCAng, tapPos))
-          .orElseGet(
-              () ->
-                  new Transformer3WResult(
-                      time, inputModel, iAMag, iAAng, iBMag, iBAng, iCMag, iCAng, tapPos));
+      return new Transformer3WResult(
+          time, inputModel, iAMag, iAAng, iBMag, iBAng, iCMag, iCAng, tapPos);
     } else throw new FactoryException("Cannot process " + entityClass.getSimpleName() + ".class.");
   }
 }