+++ src/main/java/edu/ie3/datamodel/io/factory/result/ResultEntityFactory.java
@@ -20,7 +20,6 @@ import java.time.format.DateTimeFormatter;
  */
 abstract class ResultEntityFactory<T extends ResultEntity> extends EntityFactory<T, EntityData> {
 
-  protected static final String ENTITY_UUID = "uuid";
   protected static final String TIME = "time";
   protected static final String INPUT_MODEL = "inputModel";