+++ src/main/java/edu/ie3/datamodel/io/factory/timeseries/IconTimeBasedWeatherValueFactory.java
@@ -83,17 +83,12 @@ public class IconTimeBasedWeatherValueFactory extends TimeBasedWeatherValueFacto
             "sobsrad",
             "t131m");
 
-    Set<String> minParametersWithUuid = expandSet(minParameters, UUID);
-    Set<String> allParametersWithUuid = expandSet(allParameters, UUID);
-
-    return Arrays.asList(
-        minParameters, allParameters, minParametersWithUuid, allParametersWithUuid);
+    return Arrays.asList(minParameters, allParameters);
   }
 
   @Override
   protected TimeBasedValue<WeatherValue> buildModel(TimeBasedWeatherValueData data) {
     Point coordinate = data.getCoordinate();
-    java.util.UUID uuid = data.containsKey(UUID) ? data.getUUID(UUID) : java.util.UUID.randomUUID();
     ZonedDateTime time = timeUtil.toZonedDateTime(data.getField(TIME));
     ComparableQuantity<Irradiance> directIrradiance =
         data.getQuantity(DIRECT_IRRADIANCE, PowerSystemUnits.WATT_PER_SQUAREMETRE);
@@ -111,7 +106,7 @@ public class IconTimeBasedWeatherValueFactory extends TimeBasedWeatherValueFacto
             temperature,
             windDirection,
             windVelocity);
-    return new TimeBasedValue<>(uuid, time, weatherValue);
+    return new TimeBasedValue<>(time, weatherValue);
   }
 
   /**