+++ src/main/java/edu/ie3/datamodel/io/source/DataSource.java
@@ -6,8 +6,10 @@
 package edu.ie3.datamodel.io.source;
 
 import edu.ie3.datamodel.exceptions.SourceException;
-import edu.ie3.datamodel.models.UniqueEntity;
-import java.util.*;
+import edu.ie3.datamodel.models.Entity;
+import java.util.Map;
+import java.util.Optional;
+import java.util.Set;
 import java.util.stream.Stream;
 
 /** Interface that include functionalities for data sources in the database table, csv file etc. */
@@ -19,10 +21,9 @@ public interface DataSource {
    * @param entityClass class of the source
    * @return an option for the found fields
    */
-  Optional<Set<String>> getSourceFields(Class<? extends UniqueEntity> entityClass)
-      throws SourceException;
+  Optional<Set<String>> getSourceFields(Class<? extends Entity> entityClass) throws SourceException;
 
   /** Creates a stream of maps that represent the rows in the database */
-  Stream<Map<String, String>> getSourceData(Class<? extends UniqueEntity> entityClass)
+  Stream<Map<String, String>> getSourceData(Class<? extends Entity> entityClass)
       throws SourceException;
 }