+++ src/main/java/edu/ie3/datamodel/models/timeseries/individual/TimeBasedValue.java
@@ -9,7 +9,6 @@ import edu.ie3.datamodel.models.timeseries.TimeSeriesEntry;
 import edu.ie3.datamodel.models.value.Value;
 import java.time.ZonedDateTime;
 import java.util.Objects;
-import java.util.UUID;
 
 /**
  * Describes an entity of a time series by mapping a value to a timestamp
@@ -20,13 +19,9 @@ public class TimeBasedValue<T extends Value> extends TimeSeriesEntry<T>
     implements Comparable<TimeBasedValue<? extends Value>> {
   private final ZonedDateTime time;
 
-  public TimeBasedValue(UUID uuid, ZonedDateTime time, T value) {
-    super(uuid, value);
-    this.time = time;
-  }
-
   public TimeBasedValue(ZonedDateTime time, T value) {
-    this(UUID.randomUUID(), time, value);
+    super(value);
+    this.time = time;
   }
 
   public ZonedDateTime getTime() {
@@ -49,7 +44,7 @@ public class TimeBasedValue<T extends Value> extends TimeSeriesEntry<T>
 
   @Override
   public String toString() {
-    return "TimeBasedValue{" + "uuid=" + getUuid() + ", time=" + time + ", value=" + value + '}';
+    return "TimeBasedValue{" + "time=" + time + ", value=" + value + '}';
   }
 
   @Override