+++ src/test/groovy/edu/ie3/datamodel/io/factory/input/AssetInputEntityFactoryTest.groovy
@@ -6,12 +6,12 @@
 package edu.ie3.datamodel.io.factory.input
 
 import edu.ie3.datamodel.exceptions.FactoryException
+import edu.ie3.datamodel.exceptions.NotImplementedException
 import edu.ie3.datamodel.models.OperationTime
 import edu.ie3.datamodel.models.input.AssetInput
 import edu.ie3.datamodel.models.input.OperatorInput
 import edu.ie3.datamodel.utils.Try
 import edu.ie3.test.helper.FactoryTestHelper
-import org.apache.commons.lang3.NotImplementedException
 import spock.lang.Specification
 
 import java.time.ZonedDateTime