+++ src/test/groovy/edu/ie3/datamodel/io/factory/timeseries/TimeBasedSimpleValueFactoryTest.groovy
@@ -41,14 +41,12 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
     valueClass       || expectedFields
     EnergyPriceValue || [
       [
-        TimeBasedSimpleValueFactory.UUID,
         TIME,
         PRICE
       ] as Set
     ]
     SValue           || [
       [
-        TimeBasedSimpleValueFactory.UUID,
         TIME,
         ACTIVE_POWER,
         REACTIVE_POWER
@@ -56,14 +54,12 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
     ]
     PValue || [
       [
-        TimeBasedSimpleValueFactory.UUID,
         TIME,
         ACTIVE_POWER
       ] as Set
     ]
     HeatAndSValue || [
       [
-        TimeBasedSimpleValueFactory.UUID,
         TIME,
         ACTIVE_POWER,
         REACTIVE_POWER,
@@ -72,7 +68,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
     ]
     HeatAndPValue || [
       [
-        TimeBasedSimpleValueFactory.UUID,
         TIME,
         ACTIVE_POWER,
         HEAT_DEMAND
@@ -80,7 +75,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
     ]
     HeatDemandValue || [
       [
-        TimeBasedSimpleValueFactory.UUID,
         TIME,
         HEAT_DEMAND
       ] as Set
@@ -111,7 +105,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
       "price": "52.4"
     ], EnergyPriceValue)
     def expected = new TimeBasedValue(
-        UUID.fromString("78ca078a-e6e9-4972-a58d-b2cadbc2df2c"),
         time,
         new EnergyPriceValue(Quantities.getQuantity(52.4, StandardUnits.ENERGY_PRICE))
         )
@@ -132,7 +125,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
       "heatDemand": "8.0"
     ], HeatAndSValue)
     def expected = new TimeBasedValue(
-        UUID.fromString("78ca078a-e6e9-4972-a58d-b2cadbc2df2c"),
         time,
         new HeatAndSValue(Quantities.getQuantity(500.0, StandardUnits.ACTIVE_POWER_IN), Quantities.getQuantity(165.0, StandardUnits.REACTIVE_POWER_IN), Quantities.getQuantity(8.0, StandardUnits.HEAT_DEMAND))
         )
@@ -152,7 +144,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
       "heatDemand": "8.0"
     ], HeatAndPValue)
     def expected = new TimeBasedValue(
-        UUID.fromString("78ca078a-e6e9-4972-a58d-b2cadbc2df2c"),
         time,
         new HeatAndPValue(Quantities.getQuantity(500.0, StandardUnits.ACTIVE_POWER_IN), Quantities.getQuantity(8.0, StandardUnits.HEAT_DEMAND))
         )
@@ -171,7 +162,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
       "heatDemand": "8.0"
     ], HeatDemandValue)
     def expected = new TimeBasedValue(
-        UUID.fromString("78ca078a-e6e9-4972-a58d-b2cadbc2df2c"),
         time,
         new HeatDemandValue(Quantities.getQuantity(8.0, StandardUnits.HEAT_DEMAND))
         )
@@ -191,7 +181,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
       "q": "165.0"
     ], SValue)
     def expected = new TimeBasedValue(
-        UUID.fromString("78ca078a-e6e9-4972-a58d-b2cadbc2df2c"),
         time,
         new SValue(Quantities.getQuantity(500.0, StandardUnits.ACTIVE_POWER_IN), Quantities.getQuantity(165.0, StandardUnits.REACTIVE_POWER_IN))
         )
@@ -210,7 +199,6 @@ class TimeBasedSimpleValueFactoryTest extends Specification {
       "p": "500.0"
     ], PValue)
     def expected = new TimeBasedValue(
-        UUID.fromString("78ca078a-e6e9-4972-a58d-b2cadbc2df2c"),
         time,
         new PValue(Quantities.getQuantity(500.0, StandardUnits.ACTIVE_POWER_IN))
         )