+++ src/test/groovy/edu/ie3/datamodel/io/source/csv/CsvDataSourceTest.groovy
@@ -6,22 +6,13 @@
 package edu.ie3.datamodel.io.source.csv
 
 import edu.ie3.datamodel.io.naming.FileNamingStrategy
-import edu.ie3.datamodel.models.UniqueEntity
-import edu.ie3.datamodel.models.input.NodeInput
 import edu.ie3.datamodel.models.input.system.LoadInput
-import edu.ie3.test.common.SystemParticipantTestData as sptd
 import spock.lang.Shared
 import spock.lang.Specification
 
 import java.nio.file.Path
-import java.util.concurrent.ConcurrentHashMap
-import java.util.concurrent.atomic.LongAdder
-import java.util.function.Function
-import java.util.stream.Collectors
 
 class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
-  @Shared
-  Function<Map<String, String>, String> uuidExtractor = { fieldToValues -> fieldToValues.get("uuid") }
 
   // Using a groovy bug to gain access to private methods in superclass:
   // by default, we cannot access private methods with parameters from abstract parent classes, introducing a
@@ -38,19 +29,9 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
       return super.buildFieldsToAttributes(csvRow, headline)
     }
 
-    def <T extends UniqueEntity> Set<Map<String, String>> distinctRowsWithLog(
-        Class<T> entityClass, Collection<Map<String, String>> allRows) {
-      return super.distinctRowsWithLog(allRows, uuidExtractor, entityClass.simpleName, "UUID")
-    }
-
     String[] parseCsvRow(
         String csvRow, String csvSep) {
       return super.parseCsvRow(csvRow, csvSep)
-    }
-
-    String[] oldFieldVals(
-        String csvSep, String csvRow) {
-      return super.oldFieldVals(csvSep, csvRow)
     }
   }
 
@@ -116,7 +97,7 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
       "olmcharacteristic",
       "cosPhiFixed"
     ] as String[]
-    def validCsvRow = "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8,25.0,100.0,0.95,98.0,test_bmTypeInput,50.0,25.0,olm:{(0.0,1.0)},cosPhiFixed:{(0.0,1.0)}"
+    def validCsvRow = "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8,25.0,100.0,0.95,98.0,test_bmTypeInput,50.0,25.0,\"olm:{(0.0,1.0)}\",\"cosPhiFixed:{(0.0,1.0)}\""
 
     expect:
     dummyCsvSource.buildFieldsToAttributes(validCsvRow, validHeadline) == [
@@ -130,97 +111,6 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
       uuid               : "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8",
       olmcharacteristic  : "olm:{(0.0,1.0)}",
       cosPhiFixed        : "cosPhiFixed:{(0.0,1.0)}"
-    ]
-  }
-
-  def "A CsvDataSource should be able to handle deprecated invalid csvRows correctly"() {
-    expect:
-    dummyCsvSource.oldFieldVals(csvSep, csvRow) as List == resultingArray
-
-    where:
-    csvSep | csvRow                                                                                                                                                                                                                                                                                                                                                                                                              || resultingArray
-    ","    | "4ca90220-74c2-4369-9afa-a18bf068840d,{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}},node_a,2020-03-25T15:11:31Z,2020-03-24T15:11:31Z,8f9682df-0744-4b58-a122-f0dc730f6510,true,1,1.0,Höchstspannung,380.0,olm:{(0.00,1.00)},cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}"                                                     || [
-      "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "node_a",
-      "2020-03-25T15:11:31Z",
-      "2020-03-24T15:11:31Z",
-      "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "true",
-      "1",
-      "1.0",
-      "Höchstspannung",
-      "380.0",
-      "olm:{(0.00,1.00)}",
-      "cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}"
-    ]
-    ","    | "\"4ca90220-74c2-4369-9afa-a18bf068840d\",\"{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}\",\"node_a\",\"2020-03-25T15:11:31Z\",\"2020-03-24T15:11:31Z\",\"8f9682df-0744-4b58-a122-f0dc730f6510\",\"true\",\"1\",\"1.0\",\"Höchstspannung\",\"380.0\",\"olm:{(0.00,1.00)}\",\"cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}\"" || [
-      "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "node_a",
-      "2020-03-25T15:11:31Z",
-      "2020-03-24T15:11:31Z",
-      "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "true",
-      "1",
-      "1.0",
-      "Höchstspannung",
-      "380.0",
-      "olm:{(0.00,1.00)}",
-      "cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}"
-    ]
-    ";"    | "4ca90220-74c2-4369-9afa-a18bf068840d;cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)};{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}};node_a;2020-03-25T15:11:31Z;2020-03-24T15:11:31Z;8f9682df-0744-4b58-a122-f0dc730f6510;true;1;1.0;Höchstspannung;380.0;olm:{(0.00,1.00)};cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}"            || [
-      "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}",
-      "{(0.0,1.0),(0.9,1.0),(1.2,-0.3)};{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "node_a",
-      "2020-03-25T15:11:31Z",
-      "2020-03-24T15:11:31Z",
-      "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "true",
-      "1",
-      "1.0",
-      "Höchstspannung",
-      "380.0",
-      "olm:{(0.00,1.00)}",
-      "cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}"
-    ]
-    ";"    | "\"4ca90220-74c2-4369-9afa-a18bf068840d\";\"{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}\";\"node_a\";\"2020-03-25T15:11:31Z\";\"2020-03-24T15:11:31Z\";\"8f9682df-0744-4b58-a122-f0dc730f6510\";\"true\";\"1\";\"1.0\";\"Höchstspannung\";\"380.0\";\"olm:{(0.00,1.00)}\";\"cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}\"" || [
-      "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "node_a",
-      "2020-03-25T15:11:31Z",
-      "2020-03-24T15:11:31Z",
-      "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "true",
-      "1",
-      "1.0",
-      "Höchstspannung",
-      "380.0",
-      "olm:{(0.00,1.00)}",
-      "cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}"
-    ]
-    ","    | "66275bfd-978b-4974-9f73-f270165a6351,Standard,f18a5a9b-6d45-4843-be12-be6d12de0e6b,{\"type\":\"LineString\",\"coordinates\":[[7.4116482,51.4843281],[7.4116482,51.4843281]],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}},{\"type\":\"Point\",\"coordinates\":[0.25423729,0.75409836],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:0\"}}}\""                                  || [
-      "66275bfd-978b-4974-9f73-f270165a6351",
-      "Standard",
-      "f18a5a9b-6d45-4843-be12-be6d12de0e6b",
-      "{\"type\":\"LineString\",\"coordinates\":[[7.4116482,51.4843281],[7.4116482,51.4843281]],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "{\"type\":\"Point\",\"coordinates\":[0.25423729,0.75409836],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:0\"}}}"
-    ]
-    ","    | "4ca90220-74c2-4369-9afa-a18bf068840d,{\"\"type\"\":\"\"Point\"\",\"\"coordinates\"\":[7.411111,51.492528],\"\"crs\"\":{\"\"type\"\":\"\"name\"\",\"\"properties\"\":{\"\"name\"\":\"\"EPSG:4326\"\"}}},node_a,2020-03-25T15:11:31Z,2020-03-24T15:11:31Z,8f9682df-0744-4b58-a122-f0dc730f6510,true,1,1.0,Höchstspannung,380.0,\"olm:{(0.00,1.00)}\",\"cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}\""         || [
-      "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "node_a",
-      "2020-03-25T15:11:31Z",
-      "2020-03-24T15:11:31Z",
-      "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "true",
-      "1",
-      "1.0",
-      "Höchstspannung",
-      "380.0",
-      "olm:{(0.00,1.00)}",
-      "cosPhiP:{(0.0,1.0),(0.9,1.0),(1.2,-0.3)}"
     ]
   }
 
@@ -384,107 +274,5 @@ class CsvDataSourceTest extends Specification implements CsvTestDataMeta {
     "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8;25.0;100.0;0.95;98.0;test_bmTypeInput;50.0;25.0" || "wrong separator"
     "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8,25.0,100.0,0.95,98.0,test_bmTypeInput"           || "too less columns"
     "5ebd8f7e-dedb-4017-bb86-6373c4b68eb8,25.0,100.0,0.95,98.0,test_bmTypeInput,,,,"       || "too much columns"
-  }
-
-  def "A CsvDataSource should collect be able to collect empty optionals when asked to do so"() {
-
-    given:
-    ConcurrentHashMap<Class<? extends UniqueEntity>, LongAdder> emptyCollector = new ConcurrentHashMap<>()
-    def nodeInputOptionals = [
-      Optional.of(sptd.hpInput.node),
-      Optional.empty(),
-      Optional.of(sptd.chpInput.node)
-    ]
-
-    when:
-    def resultingList = nodeInputOptionals.stream().filter(dummyCsvSource.isPresentCollectIfNot(NodeInput, emptyCollector)).collect(Collectors.toList())
-
-    then:
-    emptyCollector.size() == 1
-    emptyCollector.get(NodeInput).toInteger() == 1
-
-    resultingList.size() == 2
-    resultingList.get(0) == Optional.of(sptd.hpInput.node)
-    resultingList.get(1) == Optional.of(sptd.chpInput.node)
-  }
-
-  def "A CsvDataSource should return a given collection of csv row mappings as distinct rows collection correctly"() {
-
-    given:
-    def nodeInputRow = [
-      "uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_a",
-      "operates_until": "2020-03-25T15:11:31Z",
-      "operates_from" : "2020-03-24T15:11:31Z",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"
-    ]
-
-    when:
-    def allRows = [nodeInputRow]* noOfEntities
-    def distinctRows = dummyCsvSource.distinctRowsWithLog(allRows, uuidExtractor, NodeInput.simpleName, "UUID").getOrThrow()
-
-    then:
-    distinctRows.size() == distinctSize
-    distinctRows[0] == firstElement
-
-    where:
-    noOfEntities || distinctSize || firstElement
-    0            || 0            || null
-    10           || 1            || ["uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_a",
-      "operates_until": "2020-03-25T15:11:31Z",
-      "operates_from" : "2020-03-24T15:11:31Z",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"]
-  }
-
-  def "A CsvDataSource should return a failure if the provided collection of mappings contains duplicated UUIDs with different data"() {
-
-    given:
-    def nodeInputRow1 = [
-      "uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_a",
-      "operates_until": "2020-03-25T15:11:31Z",
-      "operates_from" : "2020-03-24T15:11:31Z",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"
-    ]
-    def nodeInputRow2 = [
-      "uuid"          : "4ca90220-74c2-4369-9afa-a18bf068840d",
-      "geo_position"  : "{\"type\":\"Point\",\"coordinates\":[7.411111,51.492528],\"crs\":{\"type\":\"name\",\"properties\":{\"name\":\"EPSG:4326\"}}}",
-      "id"            : "node_b",
-      "operates_until": "2020-03-25T15:11:31Z",
-      "operates_from" : "2020-03-24T15:11:31Z",
-      "operator"      : "8f9682df-0744-4b58-a122-f0dc730f6510",
-      "slack"         : "true",
-      "subnet"        : "1",
-      "v_target"      : "1.0",
-      "volt_lvl"      : "Höchstspannung",
-      "v_rated"       : "380"
-    ]
-
-    when:
-    def allRows = [nodeInputRow1, nodeInputRow2]* 10
-    def distinctRows = dummyCsvSource.distinctRowsWithLog(allRows, uuidExtractor, NodeInput.simpleName, "UUID")
-
-    then:
-    distinctRows.failure
-    distinctRows.exception.get().message == "'NodeInput' entities with duplicated UUID key, but different field values found! Please review the corresponding input file! Affected primary keys: 4ca90220-74c2-4369-9afa-a18bf068840d"
   }
 }